Initial refactoring to make use of Razor html templating.
Add Razor framework
Move html to cshtml for both live and historical views
Move javascript assets
Refactor Historical view to make full use of Razor

Squashed commit of the following:

commit 1a084f32785edfd4f6c8bf4cf73cc73e3d7c5a11
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Fri Jun 13 10:42:18 2025 +0100

    Refactor invocation of the historical chart to make the
    data relationship explicit between script elements

commit 8aae610a4dfef9d45f84f6c9f9568f3485b72f9e
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Thu Jun 12 18:00:40 2025 +0100

    Add basic integration test for historical chart page rendering

commit 6f6dca74aa1d9b205defbf935061c7e9324fd076
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Thu Jun 12 17:55:52 2025 +0100

    Added SSR Historical page model test

commit 42040bafdd5440c9acf26bd6e33efabb3eb911f7
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Thu Jun 12 17:42:58 2025 +0100

    Historical page now using model binding for form values.

    JavaScript refactored for simplicity assuming HTML is rendered
    now that we are SSR

commit 2be2e09fa818050ee19b8bedcb27f4f6920d6675
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Thu Jun 12 16:50:49 2025 +0100

    Convert TimeSeries message retriever to Generic

commit 4e16d2e132058f42f7cda40fdd6915c5ca9d7a80
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Thu Jun 12 16:39:56 2025 +0100

    Remove integration test for API messages. We are moving to Razor
    to do SSR. NO SCHOOL LIKE THE OLD SKOOL

commit 990c92745b7dbda99bf5647ed99ae9fb26382689
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Thu Jun 12 14:12:35 2025 +0100

    Moved historical chart to Razor serverside rendering.

commit 9f46327f42b15837ac24e57b063d24d0f08ee169
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Wed Jun 11 16:59:50 2025 +0100

    Moved all static html into Razor pages

commit 4c1b27c905077068b353aab2766a4b17db3d0ca8
Author: Neil Chambers <n3llyb0y.uk@gmail.com>
Date:   Wed Jun 11 15:31:54 2025 +0100

    Setup our webapi to use razor html templating

    - Add necessary items to the host builder
    - Include the standard directory configuration for Razor Pages
    - Add new Index landing page and nav elements

+++ src/Hearty.WebApp/MessagePersistor.cs
@@ -18,15 +18,18 @@ public class MessagePersistor(
     {
         logger.LogInformation("📂 File Processor Service is starting.");
 
-        try
+        if (!File.Exists(outputFileName))
         {
-            // Create the file if it does not exist
-            await File.WriteAllTextAsync(outputFileName, string.Empty, stoppingToken);
-        }
-        catch (IOException ex)
-        {
-            logger.LogError(ex, "💥 Failed to create file {FileName}.", outputFileName);
-            return; // Exit if we can't create the file
+            try
+            {
+                logger.LogInformation("📂 Creating file {FileName}.", outputFileName);
+                await File.WriteAllTextAsync(outputFileName, string.Empty, stoppingToken);
+            }
+            catch (IOException ex)
+            {
+                logger.LogError(ex, "💥 Failed to create file {FileName}.", outputFileName);
+                throw;
+            }
         }
 
         try
